requirements.txt
google-genai>=0.1.0
streamlit>=1.28.0
python-dotenv>=1.0.0
wave 